refactor Armstrong Numbers to convert power output to int

+++ src/lecture4_basicMaths/armstrongNumbers/Main.java
@@ -32,7 +32,7 @@ class Solution {
         int sum=0;
         for (int i = 0; i < length; i++) {
             int digit = n%10;
-            sum += Math.pow(digit, length);
+            sum += (int) Math.pow(digit, length);
             n = n/10;
         }
         return original == sum;
